src/python/projects/api/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beanie==1.28.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.6
googlesearch-python==1.2.5
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.6.0
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
redis
fastapi-cache2
fastapi-cache2[redis]
aioredis